add: emailの入力フォームとpasswordの入力フォームを共通化
