+++ front/src/components/pages/login/index.vue
@@ -13,31 +13,17 @@
           v-model="valid"
           lazy-validation
         >
-          <v-text-field
+          <EmailTextField
             v-model="loginInfo.email"
-            type="email"
             label="メールアドレス"
             placeholder="メールアドレスを入力"
-            color="blue"
-            density="comfortable"
-            variant="outlined"
-            required
-            :rules="emailRules"
-          ></v-text-field>
+          ></EmailTextField>
 
-          <v-text-field
+          <PasswordTextField
             v-model="loginInfo.password"
-            :type="isVisiblePassword ? 'text' : 'password'"
             label="パスワード"
             placeholder="半角英数字6文字以上"
-            color="blue"
-            density="comfortable"
-            variant="outlined"
-            :rules="passwordRules"
-            required
-            :append-inner-icon="isVisiblePassword ? 'mdi-eye' : 'mdi-eye-off'"
-            @click:append-inner="isVisiblePassword = !isVisiblePassword"
-          ></v-text-field>
+          ></PasswordTextField>
           <ProgressButton
             width="100%"
             color="warning"
@@ -98,6 +84,8 @@ import { useRouter } from 'vue-router'
 import { refresh } from "@/common/refresh"
 import { responsiveWidth400 } from "@/common/width"
 import ProgressButton from "@/components/shared/ProgressButton.vue"
+import EmailTextField from "@/components/shared/form/EmailTextField.vue"
+import PasswordTextField from "@/components/shared/form/PasswordTextField.vue"
 
 const userStore = useUserStore()
 const flashStore = useFlashStore()
@@ -107,20 +95,10 @@ const router = useRouter()
 const progress = ref(false)
 const errorMessages: string[] = reactive([])
 const valid = ref(true)
-const isVisiblePassword = ref(false)
 const loginInfo = reactive({
   email: "",
   password: ""
 })
-
-const emailRules = [
-  (v: string) => !!v || 'メールアドレスを入力してください',
-  (v: string) => /.+@.+\..+/.test(v) || 'メールアドレスの形式が正しくありません',
-]
-const passwordRules = [
-  (v: string) => !!v || 'パスワードを入力してください',
-  (v: string) => (v && v.length >= 6) || '6文字以上で入力してください',
-]
 
 const login = async (): Promise<void> => {
   progress.value = true
