+++ front/src/components/shared/form/BaseTextField.vue
@@ -1,0 +1,14 @@
+<template>
+  <v-text-field
+    color="blue"
+    density="comfortable"
+    variant="outlined"
+    v-bind="$attrs"
+  ></v-text-field>
+</template>
+
+<script setup lang="ts">
+</script>
+
+<style scoped>
+</style>
